1Pratyush CRAG/requirements.txt
datasets>=2.19.0,<3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
sentence-transformers>=2.6.1
scikit-learn>=1.4.0
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
llama-cpp-python>=0.2.70
tqdm>=4.66.0
